Restructure EdgeMetadata for successful authentication
