+++ io.openems.backend.uiwebsocket/src/io/openems/backend/uiwebsocket/impl/OnOpen.java
@@ -68,7 +68,8 @@ public class OnOpen implements io.openems.common.websocket.OnOpen {
 		wsData.setToken(token);
 
 		// send connection successful reply
-		AuthenticateWithSessionIdNotification notification = new AuthenticateWithSessionIdNotification(token, user);
+		AuthenticateWithSessionIdNotification notification = new AuthenticateWithSessionIdNotification(token, user,
+				User.generateEdgeMetadatas(user, this.parent.metadata));
 		this.parent.server.sendMessage(ws, notification);
 
 		this.parent.logInfo(this.log, "User [" + user.getId() + ":" + user.getName() + "] connected.");
