+++ io.openems.common/src/io/openems/common/jsonrpc/notification/AuthenticateWithSessionIdNotification.java
@@ -1,10 +1,12 @@
 package io.openems.common.jsonrpc.notification;
 
+import java.util.List;
 import java.util.UUID;
 
 import com.google.gson.JsonObject;
 
 import io.openems.common.jsonrpc.base.JsonrpcNotification;
+import io.openems.common.jsonrpc.response.AuthenticateWithPasswordResponse.EdgeMetadata;
 import io.openems.common.session.AbstractUser;
 import io.openems.common.utils.JsonUtils;
 
@@ -12,18 +14,14 @@ import io.openems.common.utils.JsonUtils;
  * Represents a JSON-RPC Notification for UI authentication with a `session_id`
  * or `token`.
  * 
- * <p>
- * 
- * See {@link AbstractUser#toJsonObject()} for details.
- * 
  * <pre>
  * {
  *   "jsonrpc": "2.0",
- *   "method": "authenticatedWithToken",
+ *   "method": "authenticatedWithSessionId",
  *   "params": {
  *     "token": String,
- *     "user": {}
- *     "edges": {}
+ *     "user": {@link AbstractUser#toJsonObject()}
+ *     "edges": {@link EdgeMetadata#toJson(java.util.Collection)}
  *   }
  * }
  * </pre>
@@ -34,17 +32,21 @@ public class AuthenticateWithSessionIdNotification extends JsonrpcNotification {
 
 	private final UUID token;
 	private final AbstractUser user;
+	private final List<EdgeMetadata> edges;
 
-	public AuthenticateWithSessionIdNotification(UUID token, AbstractUser user) {
+	public AuthenticateWithSessionIdNotification(UUID token, AbstractUser user, List<EdgeMetadata> edges) {
 		super(METHOD);
 		this.token = token;
 		this.user = user;
+		this.edges = edges;
 	}
 
 	@Override
 	public JsonObject getParams() {
-		return JsonUtils.buildJsonObject(this.user.toJsonObject()) //
+		return JsonUtils.buildJsonObject() //
 				.addProperty("token", this.token.toString()) //
+				.add("user", this.user.toJsonObject()) //
+				.add("edges", EdgeMetadata.toJson(this.edges)) //
 				.build();
 	}
