+++ io.openems.common/src/io/openems/common/jsonrpc/response/AuthenticateWithPasswordResponse.java
@@ -1,12 +1,16 @@
 package io.openems.common.jsonrpc.response;
 
+import java.util.Collection;
+import java.util.List;
 import java.util.UUID;
 
+import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
 
 import io.openems.common.jsonrpc.base.JsonrpcResponseSuccess;
-import io.openems.common.jsonrpc.shared.EdgeMetadata;
 import io.openems.common.session.AbstractUser;
+import io.openems.common.session.Role;
+import io.openems.common.types.SemanticVersion;
 import io.openems.common.utils.JsonUtils;
 
 /**
@@ -18,7 +22,7 @@ import io.openems.common.utils.JsonUtils;
  *   "id": "UUID",
  *   "result": {
  *     "token": String,
- *     "user": {@link UserMetadata#toJsonObject()}
+ *     "user": {@link AbstractUser#toJsonObject()}
  *     "edges": {@link EdgeMetadata#toJson(java.util.Collection)}
  *   }
  * }
@@ -26,19 +30,79 @@ import io.openems.common.utils.JsonUtils;
  */
 public class AuthenticateWithPasswordResponse extends JsonrpcResponseSuccess {
 
+	public static class EdgeMetadata {
+
+		/**
+		 * Converts a collection of EdgeMetadatas to a JsonArray.
+		 * 
+		 * <pre>
+		 * [{
+		 *   "id": String,
+		 *   "comment": String,
+		 *   "producttype: String,
+		 *   "version: String,
+		 *   "role: "admin" | "installer" | "owner" | "guest",
+		 *   "isOnline: boolean
+		 * }]
+		 * </pre>
+		 * 
+		 * @param metadatas the EdgeMetadatas
+		 * @return a JsonArray
+		 */
+		public static JsonArray toJson(Collection<EdgeMetadata> metadatas) {
+			JsonArray result = new JsonArray();
+			for (EdgeMetadata metadata : metadatas) {
+				result.add(metadata.toJsonObject());
+			}
+			return result;
+		}
+
+		private final String id;
+		private final String comment;
+		private final String producttype;
+		private final SemanticVersion version;
+		private final Role role;
+		private final boolean isOnline;
+
+		public EdgeMetadata(String id, String comment, String producttype, SemanticVersion version, Role role,
+				boolean isOnline) {
+			this.id = id;
+			this.comment = comment;
+			this.producttype = producttype;
+			this.version = version;
+			this.role = role;
+			this.isOnline = isOnline;
+		}
+
+		public JsonObject toJsonObject() {
+			return JsonUtils.buildJsonObject() //
+					.addProperty("id", this.id) //
+					.addProperty("comment", this.comment) //
+					.addProperty("producttype", this.producttype) //
+					.addProperty("version", this.version.toString()) //
+					.add("role", this.role.asJson()) //
+					.addProperty("isOnline", this.isOnline) //
+					.build();
+		}
+	}
+
 	private final UUID token;
 	private final AbstractUser user;
+	private final List<EdgeMetadata> edges;
 
-	public AuthenticateWithPasswordResponse(UUID id, UUID token, AbstractUser user) {
+	public AuthenticateWithPasswordResponse(UUID id, UUID token, AbstractUser user, List<EdgeMetadata> edges) {
 		super(id);
 		this.token = token;
 		this.user = user;
+		this.edges = edges;
 	}
 
 	@Override
 	public JsonObject getResult() {
-		return JsonUtils.buildJsonObject(this.user.toJsonObject()) //
+		return JsonUtils.buildJsonObject() //
 				.addProperty("token", this.token.toString()) //
+				.add("user", this.user.toJsonObject()) //
+				.add("edges", EdgeMetadata.toJson(this.edges)) //
 				.build();
 	}
