--- io.openems.common/src/io/openems/common/jsonrpc/shared/package-info.java
@@ -1,3 +1,0 @@
-@org.osgi.annotation.versioning.Version("1.0.0")
-@org.osgi.annotation.bundle.Export
-package io.openems.common.jsonrpc.shared;
