+++ io.openems.edge.controller.api.websocket/src/io/openems/edge/controller/api/websocket/Utils.java
@@ -1,22 +1,22 @@
 package io.openems.edge.controller.api.websocket;
 
-import java.util.ArrayList;
 import java.util.List;
 
+import com.google.common.collect.Lists;
+
 import io.openems.common.OpenemsConstants;
-import io.openems.common.jsonrpc.shared.EdgeMetadata;
+import io.openems.common.jsonrpc.response.AuthenticateWithPasswordResponse.EdgeMetadata;
 import io.openems.common.session.Role;
 
 public class Utils {
 	/**
 	 * Gets the EdgeMetadata for one Edge.
 	 * 
-	 * @param role the Role for this Edge
-	 * @return a list of EdgeMetadatas
+	 * @param role the {@link Role} for this Edge
+	 * @return a list of {@link EdgeMetadata}s
 	 */
 	public static List<EdgeMetadata> getEdgeMetadata(Role role) {
-		List<EdgeMetadata> metadatas = new ArrayList<>();
-		metadatas.add(new EdgeMetadata(//
+		return Lists.newArrayList(new EdgeMetadata(//
 				WebsocketApi.EDGE_ID, // Edge-ID
 				WebsocketApi.EDGE_COMMENT, // Comment
 				WebsocketApi.EDGE_PRODUCT_TYPE, // Product-Type
@@ -24,6 +24,5 @@ public class Utils {
 				role, // Role
 				true // Is Online
 		));
-		return metadatas;
 	}
 }
